This makes the subscriber able to receive messages from the publisher.
The subscriber maintains a list of publishers it's connected to, and
selects on them all, and if any of them say anything it generates a
Message event for the user thread.

Also: Documented the return value of cal_i.subscriber_read(), and made
the bip_subscriber_read() function adhere to it.