This commit makes Bionet-level HAB subscriptions work.
It adds an internal (to the library) record of all known HABs (since
CAL provides this for free).  When the user adds a HAB subscription,
all the HABs that are in the internal list of all known HABs and *not*
in the list of the HABS known to the user are reported to the user
(via the user's normal HAB callback).

Updates the cal-callback to manage these two lists of HABs properly,
and only call the user when a HAB they're interested in comes or goes.