This makes wildcard HAB specifiers in Bionet subscriptions work.
This required reworking the CAL Client internals to support subscriptions
that aren't for a particular peer, but that might apply to several peers.
cal_client.init() now takes a "peer_matches" function similar to the
cal_server.init()'s topic_matches.  Defaults to strcmp() if NULL,
which of course just results in exact peer matches.  The Bionet client
library supplies a peer_matches() function that understands HAB names
and wildcards.