This moved bionet-datapoint.h into bionet-resource.h
While it makes sense to have a Resource without a Datapoint in it,
it makes no sense at all to have a Datapoint without a Resource to
contain it.

The reason is that the Datapoint contains a union to hold the value,
but the data-type (which is the descriminator for the value union)
is part of the Resource metadata and lives up in the Resource.