add a new argument to the configure script (via autoconf muck) '--enable-python' and '--disable-python' to enable or disable building of the python bindings. "enable" is the default.
add a check for python and swig and complain if they are not there and you are attempting to build with them.

add a 'clean-local' target in the python bindings dir so that it gets cleaned correctly.

fix some swig warnings by taking out the global GSList *bionet_habs and making accessor functions to it.