add counter to messages. the counter uses an unused accel axis channel in the message if it is available. by default Y-axis channel is available.
HAB reports number of lost messages. If two subsequent messages are lost this is not detected by the HAB but can be seen by human eyes by looking at the counter value and see where it jumped. 

It cannot detect with more accuracy due to the fact that is a node resets the counter does also.