Follow memory allocation convention when making up HAB-Type and HAB-ID
bionet_hab_new() takes a type and id string, which may be NULL to mean
"let the bionet library pick my type based on process name and my
id based on hostname".  But it was setting the strings to statically
allocated buffers, so any subsequent call to bionet_hab_free() would
crash the process.  We hardly ever call bionet_hab_free() -- we normally
just exit -- so this bug's been lurking for a while.