add some configurability to the debian source package
The alsa-hab requires Alsa 1.0.16 or newer, it won't build with older
versions.  Normally we handle this by saying in debian/control that the
source package Build-Depends on libasound2-dev (>= 1.0.16).

This doesnt work so well, because Hardy has alsa 1.0.15, so that
Build-Depends would mean we could never build bionet2 on Hardy.

I think there's no way to do conditional build-time dependencies based
on distribution (or better yet, package availability).  I also think
there's no way to say "Build-Recommends".

Thus, this commit adds a way to customize the debian source package
like we need, by selecting different debian/control files at the time we
create the source package.  A new, non-standard debian/configure script
determines if we're running on Hardy, and if so selects a debian/control
file which does not build-depend on libasound2-dev.  If we're on any dist
other than hardy, it selects a debian/control which *does* build-depend
on libasound2-dev.

Imperfect and hacky?  Possibly.