Add missing bionet2-dev dependencies?
The bst/bst-generic package fails to build on Intrepid because although
it build-depends on bionet2-dev, it fails to link because libdns-sd.la
is not available:

<http://bioserve.colorado.edu/bst/builders/x86-64.intrepid.deb_binary/builds/13/steps/shell_4/logs/stdio>

libdns-sd.la is in libavahi-compat-libdnssd-dev.

I would have thought the control file's "Depends: ${shlibs:Depends}"
would have taken care of this, but apparently it does not.  I checked
libssl-dev, and it explicitly depends on zlib1g-dev, so... <shrug>

I'm not sure why Hardy built but Intrepid didnt, with both of them
missing this Dependency.