Teach autoconf about inb-1.1.0 and the bst ion
It accepts --with-ion={yes,no,check}.  "check" is default.

With "--with-ion=no" it behaves like before: disable ION.

With "--with-ion=check" it tries to find a version of ION installed on
the system, and tries to determine if it's INB-1.1.0 or the BST version.
It prefers INB-1.1.0 if both are installed.  If neither is found, it
disables ION (as if --with-ion=no had been specified).

With "--with-ion=yes" it behaves like "check", except it's an error if
no ION is found.

If an ION is found it defines a variable in config.h named "ION_VERSION",
which can be either "INB-1.1.0" or "BST".  So next we'll check for this
in the code and do the right thing.