Generate cached image name only after interpreting kwargs, match twisted convention.
Interpreting kwargs might change args.filter, so we shouldn't use args.filter
for generating the string until we've applied kwargs to the args structure.

Also, twisted by default passes each CGI parameter valueset as a list, rather
than just a single element (even if it's a list of just one element).  So we 
should expect args.filter to be a list (and we only care about the first element).
Therefore, our default should be a list of a single element.