Rename make-db.sh to bdm-make-db and install it on the target system.
Rename make-db.sh to bdm-make-db, since it is going to go in /usr/bin
Install bdm-make-db -> /usr/bin/bdm-make-db
Install schema -> /usr/share/bionet2/schema

Since bdm-make-db is now available for users, it takes the -h option and 
prints some helpful text.

bdm-make-db will prefer a file named "schema" in the CWD over the sitewide
schema.  This is required for the tests.  The tests are also updated to refer
to the new bdm-make-db.