increment simulate_loops even if we're looping forever
This commit fixes a bug i introduced in r1528: when loops==0 it would
bypass the "++simulate_loops < loops" check, which would also skip the
incrementing of simulate_loops. As a result it would always think it
was on the first loop, and always try to add all the new nodes.