Add test 122 which verifies that BDM sync bundle_lifetime works.
The BDM Sync will soon have an option that allows you to specify a bundle
lifetime.  If your network is disrupted for longer than this lifetime, then
sync bundles should be discarded.  Else, sync bundles should not be
discarded for lifetime exceeded.  This test exposes to both cases and verifies
that the setting has the desired intent.

Since the feature hasn't been implemented yet, this test should fail.