Add bundle_lifetime option to BDM sync sender.
In a BDM sync sender configuration file, you may now specify a
bundle_lifetime option, as documented in the example configuration.
This option is eventually passed to the API of the bundle protocol agent that
is sending sync bundles on your behalf (currently, only ION).  The bundles
carry that lifetime in their RFC5050 Primary Bundle Blocks, and other BPAs
in the network can use that lifetime to expire and remove the bundle from the
network (like IP's TTL, but wall-time instead of hopcount).

If the bundle_lifetime is not specified, a default of 300 seconds will be used.