Add a first try at BP commanding of Bionet resources
This adds two programs: the bionet-bp-command-proxy and the
bionet-bp-commander.  The bionet-bp-commander sends a bundle to the
bionet-bp-command-proxy, containing two strings: a Resource name and
a value.  The proxy uses the normal bionet client API to forward the
command to the proper HAB(s).

This commit also adds a test which demonstrates it working.