make bdm-make-db take a --quiet option
This commit switches the script to use the shell getopt utility for
command-line argument parsing, then uses that to add a -q options which
makes it not print the schema to stdout when it makes the db.

This will be useful for creating the database file in the bionet2 postinst
without spewing junk to the screen.