Make the test-pattern-hab call hab_read() in between events.
Before this commit, the test-pattern-hab would usleep() and sleep() for the difference in time between two events, and then implement the event.  It would sometimes call hab_read() for an event (a LOST_NODE), but never call during the sleep() or usleep()s.  This leads to the behavior that clients that join bionet in between two events won't get bionet traffic until after a (or possibly several) new events fire.

The observed behavior is that if you have a test pattern that publishes one datapoint and then waits a long time and closes:

2010-06-15 20:00:00.000000 + X-Acceleration
    average float Sensor ?

2010-06-15 20:00:01.000000 X-Acceleration average '1234.5341'

2010-06-15 21:00:00.000000 - X-Acceleration

then, a bionet-monitor (or bionet-watcher or bionet-get) that joins in between when the 1234.5341 datapoint is published, and when the node is destroyed, will never see the datapoint.

This commit improves that behavior by replacing usleep() and sleep() with a select loop: we ensure we spend the appropriate amount of time in select()/hab_read() (the sleep_duration), but we still call hab_read() to service new clients as needed.