Redo how our debian/configure script manages the dsc
We currently only have one choice when configuring the debian package: are
we running on Ubuntu "newer-than-Hardy" (in which case we build-depend
on libasound-dev and install alsa-hab), or are we running on Ubuntu
Hardy (in which case we don't).  Thus it was somewhat reasonable to just
maintain two copies ("hardy", and "standard" aka "newer") of the control
and bionet2-streams.install files.

I'm planning to add similar dsc-build-time control of whether or not
the dsc build-depends on ion, so with this commit i'm changing the way
these two files are managed.  The new way uses a single version of each
file (shared across all distributions and configuration choices), with
internal pattern substitution.