fix this test so it can run on hardy
Before this commit, the test called valgrind with a command-line argument
that is not supported in Hardy's version of valgrind: --track-origins=yes

This commit checks for the presence of the option in valgrind's --help
output, and only uses it if the local valgrind supports it.  So the test
will now use the feature if it's there, and run without it if not.