break conditionals and their blocks onto separate lines
When you program "if (condition) statement;", the lcov code coverage
analyzer can't easily tell you if you've tested the conditional-is-true
situation - it just tells you that you've stepped on the line containing
both the conditional evaluation and the statement.

By putting the statement in an explicit block, on its own line, the code
coverage is instantly obvious.