Add #DEBHELPER# magic to Debian install/rm scripts to fix pysupport.
Bionet2 uses dh_pysupport to install its python libraries in the right place.
But, before this commit, the bionet2-bdmplot package's python was not being
handled correctly by pysupport.  Brian tracked down the cause of this to the
fact that the bionet2-bdmplot postinst script was not calling
update_python_modules.  dh_pysupport was properly adding the call to
bionet2-bdmplot.postinst.debhelper, but because the .postinst script didn't have
#DEBHELPER# in it, the contents of .debhelper were not being foled into the
.postinsts.  This commit fixes that by adding #DEBHELPER# magic tokens to all
of the customized install/rm scripts.